refactor(user-registration): pack db related files together
